2025-May-07 c
2025-May-07 c

+++ Java/SpringBoot/hms2/src/main/java/com/example/hms2/TxnPoHeader.java
@@ -72,11 +72,11 @@ public class TxnPoHeader {
 		this.status = status;
 	}
 	
-	public float getTotalAmount() {
+	public BigDecimal getTotalAmount() {
 		return totalAmount;
 	}
 
-	public void setTotalAmount(float totalAmount) {
+	public void setTotalAmount(BigDecimal totalAmount) {
 		this.totalAmount = totalAmount;
 	}
